backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

langchain==0.1.0
langchain-groq==0.0.1
langchain-chroma==0.1.4
langchain-core==1.0.7
langchain-huggingface==0.0.1
chromadb==0.4.22
groq==0.4.1

pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

sentence-transformers==2.2.2
